Start using jekyll-multiple-languages-plugin
This plugin adds a folder for each language defined in config. As
a result we moved all posts to each language folder so that we keep the
current behavior of mixing languages.

Now, to add new pages we'll have to add the template plus its
translations. That is, about.html with a call to `translate_file`,
_i18n/ca/about.md and _i18n/es/about.md.

Then, all translation keys live in either _i18n/ca.yml and _i18n/es.yml.

See more details in:
https://github.com/Anthony-Gaudino/jekyll-multiple-languages-plugin