Add a build & deploy script
This script uses the _site/ dir to clone the deploy branch, in our case
`master`, build the Jekyll site and push the resulting files to it so that
Github pages serves them.

Commit messages will have the shape "Pages built at <source_commit_SHA>"
so that the `git log` can be used as a log to track what was deployed
when.