Use a data volume to store the node_modules
From the docker docs:

> If you start a container which creates a new volume, as above, and the
container has files or directories in the directory to be mounted (such
as /app/ above), the directory’s contents are copied into the volume.

This keeps the `node_modules` that exist already in /app/ when we mount
the current directory (.) into /app/.

This allows me to successfully run `docker-compose up` on a new and
fresh machine.